+++ yash-syntax/src/input.rs
@@ -16,9 +16,6 @@
 
 //! Methods about passing [source](crate::source) code to the [parser](crate::parser).
 
-use crate::source::lines;
-use crate::source::Code;
-use crate::source::Lines;
 use crate::source::Location;
 use crate::source::Source;
 use async_trait::async_trait;
@@ -37,7 +34,7 @@ pub struct Context;
 pub type Error = (Location, std::io::Error);
 
 /// Result of the [Input] function.
-pub type Result = std::result::Result<Code, Error>;
+pub type Result = std::result::Result<String, Error>;
 
 /// Line-oriented source code reader.
 ///
@@ -46,7 +43,7 @@ pub type Result = std::result::Result<Code, Error>;
 pub trait Input {
     /// Reads a next line of the source code.
     ///
-    /// The input function is line-oriented; that is, this function returns a [`Code`] that is
+    /// The input function is line-oriented; that is, this function returns a string that is
     /// terminated by a newline unless the end of input (EOF) is reached, in which case the
     /// remaining characters up to the EOF must be returned without a trailing newline. If there
     /// are no more characters at all, the returned line is empty.
@@ -61,25 +58,22 @@ pub trait Input {
 
 /// Input function that reads from a string in memory.
 pub struct Memory<'a> {
-    lines: Lines<'a>,
+    lines: std::str::SplitInclusive<'a, char>,
 }
 
 impl Memory<'_> {
     /// Creates a new `Memory` that reads the given string.
-    pub fn new(code: &str, source: Source) -> Memory<'_> {
-        let lines = lines(code, source);
+    pub fn new(code: &str, _source: Source) -> Memory<'_> {
+        // FIXME Remove the source parameter
+        let lines = code.split_inclusive('\n');
         Memory { lines }
-    }
-
-    fn next_line_sync(&mut self, _: &Context) -> Code {
-        self.lines.next_or_empty()
     }
 }
 
 #[async_trait(?Send)]
 impl Input for Memory<'_> {
-    async fn next_line(&mut self, context: &Context) -> Result {
-        Ok(self.next_line_sync(context))
+    async fn next_line(&mut self, _context: &Context) -> Result {
+        Ok(self.lines.next().unwrap_or("").to_owned())
     }
 }
 
@@ -94,9 +88,7 @@ mod tests {
         let mut input = Memory::new("", Source::Unknown);
 
         let line = block_on(input.next_line(&Context)).unwrap();
-        assert_eq!(*line.value.borrow(), "");
-        assert_eq!(line.start_line_number.get(), 1);
-        assert_eq!(line.source, Source::Unknown);
+        assert_eq!(line, "");
     }
 
     #[test]
@@ -104,14 +96,10 @@ mod tests {
         let mut input = Memory::new("one\n", Source::Unknown);
 
         let line = block_on(input.next_line(&Context)).unwrap();
-        assert_eq!(*line.value.borrow(), "one\n");
-        assert_eq!(line.start_line_number.get(), 1);
-        assert_eq!(line.source, Source::Unknown);
+        assert_eq!(line, "one\n");
 
         let line = block_on(input.next_line(&Context)).unwrap();
-        assert_eq!(*line.value.borrow(), "");
-        assert_eq!(line.start_line_number.get(), 2);
-        assert_eq!(line.source, Source::Unknown);
+        assert_eq!(line, "");
     }
 
     #[test]
@@ -119,23 +107,15 @@ mod tests {
         let mut input = Memory::new("one\ntwo\nthree", Source::Unknown);
 
         let line = block_on(input.next_line(&Context)).unwrap();
-        assert_eq!(*line.value.borrow(), "one\n");
-        assert_eq!(line.start_line_number.get(), 1);
-        assert_eq!(line.source, Source::Unknown);
+        assert_eq!(line, "one\n");
 
         let line = block_on(input.next_line(&Context)).unwrap();
-        assert_eq!(*line.value.borrow(), "two\n");
-        assert_eq!(line.start_line_number.get(), 2);
-        assert_eq!(line.source, Source::Unknown);
+        assert_eq!(line, "two\n");
 
         let line = block_on(input.next_line(&Context)).unwrap();
-        assert_eq!(*line.value.borrow(), "three");
-        assert_eq!(line.start_line_number.get(), 3);
-        assert_eq!(line.source, Source::Unknown);
+        assert_eq!(line, "three");
 
         let line = block_on(input.next_line(&Context)).unwrap();
-        assert_eq!(*line.value.borrow(), "");
-        assert_eq!(line.start_line_number.get(), 3);
-        assert_eq!(line.source, Source::Unknown);
+        assert_eq!(line, "");
     }
 }