+++ yash-syntax/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `Annotation`'s field `location` changed to a reference
     - `Annotation`'s field `code` added
     - `Annotation`'s method `new` added
+- Items in the `input` module:
+    - `Result` redefined as `Result<String, Error>` (previously `Result<Code, Error>`)
 - `Lexer::new` now requiring the `start_line_number` and `source` parameters
 - Dependency versions
     - `async-trait` 0.1.50 → 0.1.52